moved ravel_pytree to einstein/util.py and renamed to batch_ravel_pytree. changed batch_ravel_pytree to only flatten batch dimensions to single dimension.
