+++ test/test_util.py
@@ -6,19 +6,14 @@ from numpy.testing import assert_allclose
 import pytest
 
 from jax import random
+from jax.flatten_util import ravel_pytree
 import jax.numpy as jnp
 from jax.test_util import check_eq
 from jax.tree_util import tree_flatten, tree_multimap
 
 import numpyro
 import numpyro.distributions as dist
-from numpyro.util import (
-    check_model_guide_match,
-    fori_collect,
-    format_shapes,
-    ravel_pytree,
-    soft_vmap,
-)
+from numpyro.util import check_model_guide_match, fori_collect, format_shapes, soft_vmap
 
 
 def test_fori_collect_thinning():
@@ -86,30 +81,6 @@ def test_fori_collect_return_last(progbar):
 )
 def test_ravel_pytree(pytree):
     flat, unravel_fn = ravel_pytree(pytree)
-    unravel = unravel_fn(flat)
-    tree_flatten(tree_multimap(lambda x, y: assert_allclose(x, y), unravel, pytree))
-    assert all(
-        tree_flatten(
-            tree_multimap(
-                lambda x, y: jnp.result_type(x) == jnp.result_type(y), unravel, pytree
-            )
-        )[0]
-    )
-
-
-@pytest.mark.parametrize(
-    "pytree",
-    [
-        {
-            "a": np.array([[1.0, 0.0], [1, 0.0]]),
-            "b": np.array([[1.0, 2.0], [3.0, 4.0]]),
-        },
-    ],
-)
-@pytest.mark.parametrize("nbatch_dims", [0, 1, 2])
-def test_ravel_pytree_batched(pytree, nbatch_dims):
-    flat, unravel_fn = ravel_pytree(pytree, batch_dims=nbatch_dims)
-    assert jnp.ndim(flat) == nbatch_dims + 1
     unravel = unravel_fn(flat)
     tree_flatten(tree_multimap(lambda x, y: assert_allclose(x, y), unravel, pytree))
     assert all(
